requirements-dev.txt
-r requirements.txt
ruff==0.13.0
pytest-cov==7.0.0